refactor: 결제 DTO를 컨트롤러에서 도메인 패키지로 이동
