+++ src/main/java/com/ururulab/ururu/payment/dto/request/PaymentConfirmRequestDto.java
@@ -1,6 +1,6 @@
-package com.ururulab.ururu.payment.controller.dto.request;
+package com.ururulab.ururu.payment.dto.request;
 
-import com.ururulab.ururu.payment.controller.dto.validation.PaymentValidationMessages;
+import com.ururulab.ururu.payment.dto.validation.PaymentValidationMessages;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
