+++ src/main/java/com/ururulab/ururu/payment/dto/response/PaymentConfirmResponseDto.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.payment.controller.dto.response;
+package com.ururulab.ururu.payment.dto.response;
 
 import com.ururulab.ururu.payment.domain.entity.enumerated.PaymentStatus;
