+++ src/main/java/com/ururulab/ururu/payment/service/PaymentService.java
@@ -12,13 +12,13 @@
     import com.ururulab.ururu.order.domain.repository.CartRepository;
     import com.ururulab.ururu.order.domain.repository.OrderRepository;
     import com.ururulab.ururu.order.service.StockReservationService;
-    import com.ururulab.ururu.payment.controller.dto.request.PaymentConfirmRequestDto;
-    import com.ururulab.ururu.payment.controller.dto.request.PaymentRequestDto;
-    import com.ururulab.ururu.payment.controller.dto.request.TossWebhookDto;
-    import com.ururulab.ururu.payment.controller.dto.response.PaymentConfirmResponseDto;
-    import com.ururulab.ururu.payment.controller.dto.response.PaymentFailResponseDto;
-    import com.ururulab.ururu.payment.controller.dto.response.PaymentResponseDto;
-    import com.ururulab.ururu.payment.controller.dto.response.TossPaymentResponseDto;
+    import com.ururulab.ururu.payment.dto.request.PaymentConfirmRequestDto;
+    import com.ururulab.ururu.payment.dto.request.PaymentRequestDto;
+    import com.ururulab.ururu.payment.dto.request.TossWebhookDto;
+    import com.ururulab.ururu.payment.dto.response.PaymentConfirmResponseDto;
+    import com.ururulab.ururu.payment.dto.response.PaymentFailResponseDto;
+    import com.ururulab.ururu.payment.dto.response.PaymentResponseDto;
+    import com.ururulab.ururu.payment.dto.response.TossPaymentResponseDto;
     import com.ururulab.ururu.payment.domain.entity.Payment;
     import com.ururulab.ururu.payment.domain.entity.PointTransaction;
     import com.ururulab.ururu.payment.domain.entity.enumerated.PayMethod;
