+++ src/test/java/com/ururulab/ururu/payment/service/PaymentServiceTest.java
@@ -17,12 +17,12 @@ import com.ururulab.ururu.order.domain.entity.enumerated.OrderStatus;
 import com.ururulab.ururu.order.domain.repository.CartRepository;
 import com.ururulab.ururu.order.domain.repository.OrderRepository;
 import com.ururulab.ururu.order.service.StockReservationService;
-import com.ururulab.ururu.payment.controller.dto.request.PaymentConfirmRequestDto;
-import com.ururulab.ururu.payment.controller.dto.request.PaymentRequestDto;
-import com.ururulab.ururu.payment.controller.dto.response.PaymentConfirmResponseDto;
-import com.ururulab.ururu.payment.controller.dto.response.PaymentFailResponseDto;
-import com.ururulab.ururu.payment.controller.dto.response.PaymentResponseDto;
-import com.ururulab.ururu.payment.controller.dto.response.TossPaymentResponseDto;
+import com.ururulab.ururu.payment.dto.request.PaymentConfirmRequestDto;
+import com.ururulab.ururu.payment.dto.request.PaymentRequestDto;
+import com.ururulab.ururu.payment.dto.response.PaymentConfirmResponseDto;
+import com.ururulab.ururu.payment.dto.response.PaymentFailResponseDto;
+import com.ururulab.ururu.payment.dto.response.PaymentResponseDto;
+import com.ururulab.ururu.payment.dto.response.TossPaymentResponseDto;
 import com.ururulab.ururu.payment.domain.entity.Payment;
 import com.ururulab.ururu.payment.domain.entity.enumerated.PaymentStatus;
 import com.ururulab.ururu.payment.domain.repository.PaymentRepository;
